Rename StructuredPopup
